+++ tibiaapi/Packets/Pipes/UpdateCreatureTextPacket.cs
@@ -7,7 +7,7 @@ namespace Tibia.Packets.Pipes
 {
     public class UpdateCreatureTextPacket : PipePacket
     {
-        public int CreatureId{get;set;}
+        public uint CreatureId{get;set;}
         public string CreatureName{get;set;}
         public Location Location{get;set;}
         public string Text{get;set;}
@@ -25,7 +25,7 @@ namespace Tibia.Packets.Pipes
 
             Type = PipePacketType.UpdateCreatureText;
             
-            CreatureId = (int)msg.GetUInt32();
+            CreatureId = msg.GetUInt32();
             CreatureName = msg.GetString();
             Location = new Location((int)msg.GetUInt32(), (int)msg.GetUInt32(), 0);
             Text = msg.GetString();
@@ -38,7 +38,7 @@ namespace Tibia.Packets.Pipes
             NetworkMessage msg = NetworkMessage.CreateUnencrypted(Client, 13 + CreatureName.Length + Text.Length);
             msg.AddByte((byte)Type);
 
-            msg.AddUInt32((uint)CreatureId);
+            msg.AddUInt32(CreatureId);
             msg.AddString(CreatureName);
             msg.AddUInt16((ushort)Location.X);
             msg.AddUInt16((ushort)Location.Y);
@@ -47,7 +47,7 @@ namespace Tibia.Packets.Pipes
             return msg.Data;
         }
 
-        public static bool Send(Objects.Client client, int creatureId , string creatureName, Location location, string text)
+        public static bool Send(Objects.Client client, uint creatureId , string creatureName, Location location, string text)
         {
             UpdateCreatureTextPacket p = new UpdateCreatureTextPacket(client);
